[REFACTOR] Data Class

+++ app/src/main/java/com/android/wearther/data/model/current/Coord.kt
@@ -1,6 +1,6 @@
 package com.android.wearther.data.model.current
 
 data class Coord(
-    val lat: Int,
+    val lat: Double,
     val lon: Double
 )
